Extract Move class into his own module
