+++ src/game/Move.py
@@ -1,0 +1,5 @@
+class Move():
+    TOP = 'TOP'
+    RIGHT = 'RIGHT'
+    BOTTOM = 'BOTTOM'
+    LEFT = 'LEFT'
